requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.25.0
rich>=13.7.0
flask>=3.0.0
openai>=1.12.0